refactor: Make the gatsby build-side code modular, so it's clear what the high-level steps being taken are.
